server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
sqlalchemy==1.4.23
werkzeug==2.0.1
pyjwt[crypto]==2.10.1
pillow==8.3.1
numpy==1.21.2
opencv-python==4.5.3.56
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.2.3
Flask-SQLAlchemy==2.5.1
requests==2.31.0